// src/models/Person.ts
export interface Person {
    id:         number;
    nombres:           string;
    apellidoPaterno:   string;
    apellidoMaterno:   string;
    numeroDocumento:   string;
    razonSocial:       string;
    idTipoPersona:     string;
    tipoPersona:       string;
    correoElectronico: string;
    telefono:          string;
    activo:            boolean;
    idUsuarioRegistro: string;
    idUsuarioEdicion:  string;
}