// src/models/LogisticDetail.ts
export interface LogisticDetail {
    idLinea:              number;
    idProducto:           number;
    descripcionExtendida: string;
    codigo?:               string;
    unidad?:               string;
    marca?:                string;
    modelo?:               string;
    cantidad:             number;
}